# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarberShop API",
        "description": "Appointment Scheduling System for Barbershops.",
        "contact": {
            "url": "https://github.com/matheusmrqs4"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/appointment/{appointment}/evaluate": {
            "post": {
                "tags": [
                    "Evaluation"
                ],
                "summary": "Submit a new evaluation",
                "operationId": "01f4a43ea148d9ffe085c76b46c3a305",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointments_id",
                                    "comment",
                                    "grade"
                                ],
                                "properties": {
                                    "appointments_id": {
                                        "description": "ID of the appointment",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "comment": {
                                        "description": "Comment for the evaluation",
                                        "type": "string",
                                        "example": "Great service!"
                                    },
                                    "grade": {
                                        "description": "Grade given for the appointment",
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evaluation submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Evaluation submitted successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid data provided"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - User is not authorized to evaluate this appointment"
                    },
                    "404": {
                        "description": "Not Found - Appointment not found"
                    },
                    "500": {
                        "description": "Error submitting evaluation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/evaluation/{evaluation}": {
            "get": {
                "tags": [
                    "Evaluation"
                ],
                "summary": "Get details of a specific evaluation",
                "operationId": "f2a28d8fe7f03803b7b4b19ff8214c15",
                "parameters": [
                    {
                        "name": "evaluation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evaluation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - Evaluation not found"
                    },
                    "500": {
                        "description": "Error fetching evaluation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}